backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
python-dotenv
reportlab
email-validator
